refactor: separate modules into files
